+++ services/iaas/CHANGELOG.md
@@ -20,6 +20,9 @@
   - Model `CreateServerPayloadBootVolume` of `BootVolume` property changed to `ServerBootVolume`
   - Property `Networking` in `CreateServerPayload` is required now
 
+## v0.3.1
+- Bump dependency `cloud.stackit.sdk.core` to v0.4.1
+
 ## v0.3.0
 - **Feature:** Add `createdAt` and `updatedAt` attributes to `SecurityGroupRule`, `BaseSecurityGroupRule`, `CreateSecurityGroupRulePayload` model classes
 - **Feature:** Add `description` attribute to `CreateNicPayload`, `NIC`, `UpdateNicPayload` model classes